scripts/requirements.txt
PyPDF2>=3.0.0
cohere>=4.37
python-dotenv>=1.0.0
requests>=2.31.0
pymongo>=4.6.0
motor>=3.3.0
